build_debs.sh:  Partial fix for building arch independent packages.
We only want to build the architecture independent packages once.
The i386 platform has been picked for that, and by convention there
should only be one i386 host in the deb_hosts list.  For this host
the -b option is supplied in debbuildopts which builds the binary
(No source) version of arch dependent and independent packages.  For
all other hosts the pbuilder --binary-arch flag is supplied which
builds binary arch dependent packages only.