build:  rysync debian build results into seperate directories
Avoid collisions of files with the same name, but different checksums,
by keeping the results from different build platforms seperate.  This
wasn't a problem before because these files were not modified during the
build.  With the addition of the all arch and source builds on the i386
platform, some files are now changed (bug #3969).

Signed-off-by: Gary Lowell  <gary.lowell@inktank.com>