jenkins: use in-tree clone for the ceph-build repo
Historically we have had a single system-wide clone of the ceph-build
repository on each builder, at /srv/ceph-build.

The disadvantage to this configuration is that there is no way to ensure
that each of these clones are up to date. When we make changes to the
ceph-build repository on GitHub, we have to manually pull the changes
into each builder's local clone at /srv.

Going forward, we will use the Multi-SCM plugin to Jenkins to make the
ceph-setup Jenkins job use an up-to-date clone of the ceph-build
repository. This will ensure that Jenkins is using the very latest code
from ceph-build.

The ceph-build repository will be cloned into the root of the tree.
Remove the references to /srv/ceph-build and just use the relative path,
"ceph-build".