release_tarball: select the "v" tags for cephver
By default, Jenkins' Git plugin adds its own tag to the repository. This
interferes with the versioning in release_tarball.sh

One solution would be to disable the tag in the Jenkins job. Since this
is not the default, that tag creation option would need to be manually
disabled for every job in Jenkins that runs this script. The other
option is to make the script more selective about what it considers to
be cephver. This commit implements this option.

Use a --match selector to "git describe" so that we're sure we're using
the standard "vX.X" tags.

Pipe "git describe" through "sed" instead of "cut", so that it's easier
to debug what tag was selected.

Signed-off-by: Ken Dreyer <ken.dreyer@inktank.com>