unquote archive settings
Prior to this commit the archive settings for ceph-deploy and
radosgw-agent were quoted 'false' or 'true' instead of simply the
boolean types that YAML expects.

jenkins-jobs interpreted these strings to be "true", so we ended up with
the following XML:

  <latestOnly>true</latestOnly>

This meant that Jenkins deleted all the older archived ceph-deploy
packages each time a new version was built.  When 1.5.22.1 was built,
the "older" 1.5.23 package was deleted.

Unquote the boolean values so JJB will properly interpret them.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>