remove fedora20 Jenkins slave
Remove the Fedora 20 build slave from the job configs.

We're having trouble resurrecting
gitbuilder-cdep-rpm-cloud-fedora20-amd64-basic.cloud.sepia.ceph.com in
DreamCompute, and Fedora 20 itself is going EOL in a month or so.

This is the only Fedora builder we have, and this VM does build packages
for http://ceph.com/rpm-hammer/fc20/, so it would be nice to get fresh
Fedora build slaves set up eventually. But in the meantime, since this
one is giving us trouble, just remove it.