add ceph-docs job
This job publishes ceph's in-tree documentation to the docs.ceph.com web
server.

Gitbuilder previously performed this function, so we can retire the
gitbuilder host (gitbuilder-doc.front.sepia.ceph.com) that did this,
once we're satisfied that this job is running smoothly.

TODO: Replace "newdocs.ceph.com" with "docs.ceph.com" in the rsync
commands and SSH host key. This can be done once we change the
"docs.ceph.com" A record in DNS to point to the the newdocs.ceph.com web
server.