ansible: use "ceph-jenkins" as API user
Jenkins authorizes this user account to add or remove slaves. In order
to have permission to add or remove slaves, we have to give this account
administrator rights to Jenkins.

Prior to this commit, we had a system user "prado" in Jenkins with
administrator rights.

Since we do not control this user account on GitHub.com, it is better to
use a GitHub.com user account that we control.