ansible: do not pre-populate an SSH key for docs.ceph.com
The new docs VM is moving behind a firewall / reverse-proxy, so it will
not be directly accessible from the internet. We will to run the
documentation jobs directly on the docs VM as a Jenkins slave, and we
will stop using rsync over SSH for publishing the docs.

Since there's no direct SSH access to the docs VM any more, remove the
SSH host key from the slaves.

(My Ansible syntax in slave.yml.j2 was broken anyway.)