jenkins-job-builder: add FORCE parameter
This will wipe out JJB's cache if set. It is useful for wiping out
changes that may have occurred outside of GitHub's master branch.

For example, if I make a change to a job on a local branch, and I apply
my JJB changes locally, I want to force this jenkins-job-builder job to
wipe out my changes later.