rm date/time tracking
Prior to this commit, some of the build scripts would print the time and
date to the console log.

As far as I can tell this is a hold-over from the original Gitbuilder
scripts. Jenkins already logs the start and end times for us so there's
no need to clutter the build scripts with this.

Remove the code that prints `date`.