make ceph-ansible-pull-requests request nodes specific to the job
When the labels were more generic other jobs would pick them up and run
the risk of being destroyed by the ceph-ansible-pull-requests job after
a run. There was also an issue were jobs that didn't need a node with an
extra device would provision one of the 'unique' nodes because the
labels also included 'small' and 'trusty'. Making the labels this
specific ensures that only this job will use those node types.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>