ceph-pull-requests: Override processor numbers
By default, run-make-check divide by 2 the number of processors to be used
during the make process.

When building a PR on our build farm, we need to use the full potential of the
build hosts.

By overriding NPROC with the actual number of processors, we speedup the build
process and use in a more efficient way the build hosts.