ceph-build-os: Adding ceph-build-os job
This new job is about building VM image intented to build Ceph base code.

It uses the edeploy project to define the content of the VM by defining :
- the distro
- the distro version
- a release version of that build

The ceph-builder target of edeploy intent to create xfs-based VMs that have all
the usual packages required to build ceph or edeploy.

This will save time at build-time as most of the deps will be already installed.