ceph-pull-requests: Increasing ccache size
By default, ccache is now configured and uses only 1G of space.
While monitoring the build it appears that if we would have a larger cache, more
hits could be done making a faster build process.

This patch simply makes the ccache bigger up to 5G which is fine regarding the
openstack flavor we have.