ceph-build: remove the --binary-arch flag when calling pbuilder
This caused for non-binary arch to be skipped. Specifically:
"Build-Depends-Indep" which in this case is in charge of building
libcephfs-java.

The man page explains that the flag is used to: "Specify  to  build
architecture specific targets instead of all targets" which we need to.

This was introduced in commit: d80f78adcb28a2c60fd79895f4f5c63d999b457d

Signed-off-by: Alfredo Deza <adeza@redhat.com>