ansible: always add 'ceph-builders' to /etc/hosts
The nodes we spin up in OVH are using an image that sets the hostname
to 'ceph-builders'. Also, the hostname we give it is too long to use and
ansible silenty fails to set it.

This change ensures that nodes can use sudo correctly and avoids errors
like: 'sudo: unable to resolve host ceph-builders'.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>