add ceph-ansible-rpm job
This job builds RPMs for each ceph-ansible GitHub branch.

(This is a 99% copy and paste from ceph-installer-rpm, so we probably
need to come up with a centralized area for these types of jobs and use
JJB templates. I can't come up with a good directory name at the moment,
though.)