ceph-*: fix tcmalloc args for pre-cmake builds
Prior to cmake, --with(out)-tcmalloc must be passed to the
configure done during the actual build.  For rpm, that means an
extra arg to the invocation of rpmbuild in build_rpm; for deb,
debian/rules invokes configure.  Also, no reason to export
the shell variables; they're used locally and passed as
property files between Jenkins jobs.

Fixes: http://tracker.ceph.com/issues/18744
Signed-off-by: Dan Mick <dan.mick@redhat.com>