ceph-pr-submodules: explicitly fetch the target branch
This ceph-pr-submodules job is still giving us false positives in some
cases.

We were comparing against a target branch (for example "origin/jewel"),
but we could not see the *exact* sha1 that Jenkins thought was
"origin/jewel". Sometimes this job would fail, and the failure hinted
that Jenkins might be comparing an older target branch.

Explicitly fetch the target branch immediately before doing the "git
log" operation, and print the target branch's sha1 ref to stdout in case
we need further debugging.