ceph-docker-lint: do not check variables_entrypoint.sh
Shellcheck will complain with SC2034: foo appears unused. Verify it or
export it.
The problem here is that this file contains a reference of all the
variables so indeed they appear unused.
We could potentially source some portion of the code from that file,
however I don't want to change the structure of the code because of
shellchecker. Also sourcing all the variables might results in expected
scenario if these variables are not expected to be declared.

There is no way to tell shellcheck to ignore this file so we should not
test it.

Signed-off-by: Sébastien Han <seb@redhat.com>