ceph-ansible-pr-syntax-check: add library path
Explicitly set the LIBRARY path to avoid the following Ansible error:

ERROR! no action detected in task. This often indicates a misspelled
module name, or incorrect module path.

We need this change because roles path changed from ceph-ansible/ to
ceph-ansible/infrastructure-playbooks so now when we are running the
syntax check it looks in ceph-ansible/infrastructure-playbooks/library
instead of ceph-ansible/library.

Signed-off-by: Sébastien Han <seb@redhat.com>