Abort builds on update
When a PR is updated, jobs that were launched by the previous PR are not
aborted. It can even happen that you update a PR several times in a
short time. So it's n*x jenkins slave locked until they end.

Adding `cancel-builds-on-update` to the jobs templates force the job to be
aborted to free a jenkins slave quicker.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>