ceph-ansible-prs: test against ansible2.3 and luminous only
This drops support for testing jewel on PRs and changes the ansible
version to 2.3. If we want to continue testing against jewel we can
schedule nightly tests that run against master. With this approach we
can still get the same coverage and reduce a lot of demand on the CI
caused by PR testing.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>