kernel: use raw xml for configuring scm module
kernel-trigger job needs to specify do-not-fetch-tags, otherwise
cloning fails, "No changes" is reported and nothing gets triggered (see
commit b34f8fa ("kernel: do not fetch git tags") for more details).
do-not-fetch-tags was added to jenkins-job-builder in commit
badab0264717 ("Add do-not-fetch-tags to CloneOption for Git"), which
isn't in any tag.  The version we get from pypi, 2.0.0.0b2, simply
ignores it, leaving us with no kernel builds.

Use raw xml generated with jenkins-job-builder's master to get
<noTags>true</noTags> up there after all.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>