ceph-container-flake8: Using project's tox.ini
The current code is running flake8 from a container to check every
python file changed by a PR.

The actual code doesn't import the tox.ini of ceph-container which
specify a particular configuration like "max-line-length = 99"

This lead to a situation where a dev which runs flake on his dev env
have a flake reporting now failures while flake from the CI fails.

So this patch is about adding project's tox.ini inside the container to
insure flake is using it.