ceph-container: add a nightly flag
In addition to
ceph/ceph-container@fa0a1483d59c5a30e99f379e12a24239f91e7980

Adding this environment variable allows tests/tox.sh in ceph/ceph-container to
know whether we are running in a nightly job.

This is needed because ceph-container project checks if some code has
been changed, if no, test is not run.
Since in nightlies job will always enter this condition we must have a
way to be aware if we are running a nightly so we can skip this test.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>