ceph-dashboard-pull-requests: Temporary remove Ubuntu node(s)
In order to be able to get to the step where the dashboard end-to-end tests
are being executed (which seem to pass on centos) the proposal is to temporary remove
the Ubuntu nodes.
When the current issues with Ubuntu have been fixed the plan is to
re-add Ubuntu (xenial) again to the configuration of this Jenkins job.
Since the job is being triggered automatically it would also prevent failing
jobs (which currently happens all the time and is eating up resources).

Signed-off-by: Laura Paduano <lpaduano@suse.com>