ceph-ansible: remove useless command for syntax
Because we're now able to execute the infrastructure playbooks from
the main ceph-ansible directory, we don't need to define the ansible
library path (because already define in ansible.cfg).
The roles and group_vars directories are already present in the main
directory so we don't need to copy them anymore.
This commit also exclude both ansible-lint rules: 204 and 205

- [204] Lines should be no longer than 160 chars
- [205] Use ".yml" or ".yaml" playbook extension

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>