slave.yml: Add jenkins-build user to jenkins-build group
On ubuntu, at least, with the addition of the "create a jenkins-build group" task, the creation of the user wasn't adding the user to the group.  I discovered this by seeing a docker command fail after the socket was 'chgrp'ed to jenkins-build

```
+ sudo chgrp jenkins-build /var/run/docker.sock
+ do_clean
+ make clean.all
Got permission denied while trying to connect to the Docker daemon socket at unix:///var/run/docker.sock
```

I logged in to the slave and confirmed this was the case.

Signed-off-by: David Galloway <dgallowa@redhat.com>