ansible: Just use the example
In our setup, the example is the canonical playbook that works.  If, for some reason, ceph-build.git is used by someone else, they can rm the symlink and create their own `ansible/slave.yml` but for our purposes, there's no need to keep separate playbooks as it just creates confusion when modifying the playbooks.

Signed-off-by: David Galloway <dgallowa@redhat.com>