ceph-dashboard-* jobs: Install Prometheus and Alertmanager
In order to test the alert configuration within the dashboard
(e2e tests) we need to have Prometheus and Alertmanager running
on the system. Added a new script which takes care of the installation
and running Prometheus and Alertmanager in the background.
Added the script only to the dashboard e2e Jenkins job (for the time being),
the processes will be killed at the end of the job.

Signed-off-by: Laura Paduano <lpaduano@suse.com>