ceph*build: Use new gigantic label
By requiring 'braggi', none of the ARM64 builds would start.  It wouldn't be appropriate to give the arm64 builders a "braggi" label so I'm creating a new "gigantic" label that can apply to the braggi and arm64 builders.

There are no nodes in mita that have a 'gigantic' label so that guarantees builds won't run on OVH ephemeral nodes.

Signed-off-by: David Galloway <dgallowa@redhat.com>