teuthology: always test pull requests in sepia
Because teuthology has functional and integration tests which
needs a connection to some services running in sepia and
some of the jenkins workers are running far away from the lab,
it takes significantly more time to run tests when using those
nodes. Like 6-7 minutes for running test on nodes in sepia,
and about 55-60 minutes outside.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>