ansible: One playbook to rule them all
I/we got into a bad habit of updating one playbook to support X new distro or X new package but wouldn't update the others.

I think having all the tasks for all the types of slaves in one playbook will help keep things homogenous.  And the playbooks are still idempotent (they can't be run while a job is running of course).

Signed-off-by: David Galloway <dgallowa@redhat.com>