kernel: drop centos7 builds
The gcc version requirement was raised to 4.9 in kernel 4.8.  centos7
is on gcc 4.8 and these builds have been failing for the last couple of
months.

While we could switch to devtoolset-7 or devtoolset-8, both of which
are installed on our centos7 build agents, let's just drop them.  For
one-off testing of older releases, trusty and xenial would do.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>