Grab SHA1 into parent job's environment for human or program
The Ceph SHA1 is divined by the -setup job, and not available in the
parent job.  That's easily fixed by the same method it's passed to the
-build job: grab the file artifact and use it to inject SHA1 into the
environment.  This means the parent job can show the SHA1 in its build
label for humans, and can also be searched by Jenkins API calls.

Signed-off-by: Dan Mick <dmick@redhat.com>