ceph-dev, ceph-dev-new: copyartifact permission: no spaces
I don't see why from examining the plugin source, but it appears
that if you use both a comma and a space, the second entry will not
be found (as though the space is becoming part of the string for
comparison inside the plugin).  Stopping fighting and just letting it happen.

Signed-off-by: Dan Mick <dmick@redhat.com>