ceph*build: Login to docker.io if needed
We ran into an issue today where the internal mirror was temporarily unavailable.  The job fell back to docker.io and despite the fact that we're not using it all that much, we still got rate limited.  I suspect other tenants in the Community Cage are still pulling from docker.io anonymously.

This PR will have us login (with a higher rate-limit).

Signed-off-by: David Galloway <dgallowa@redhat.com>