prune-quay.py: fix deletion of busy ref names (master etc.)
We must only delete the ref name if it refers to the same image
as the ref-shortsha1-el form of the name; if it's been pushed as a
new build of ref from a new sha1, it's no longer the image we
need to delete.  Verify the ref tag points to the same image as the
ref-shortsha1-el tag.

Signed-off-by: Dan Mick <dmick@redhat.com>