ceph-dev-*-setup: use libc allocator when building quincy on bionic
see also https://tracker.ceph.com/issues/39703, to silence the warning
std::allocator<void> is deprecated in C++17 and will be removed in
C++20, so stop using allocator<void> in quincy and up.

but the libtcmalloc shipped by bionic is still buggy (v2.5), so we should not
build qunincy and up with tcmalloc on this distro.

after this change, we only build quincy and up with tcmalloc on non-bionic distro.

Signed-off-by: Kefu Chai <kchai@redhat.com>