ansible: Pin vagrant-libvirt to 0.3.0
The latest version (and possibly older) 0.5.3 is giving us tons of issues.

- VMs hanging on `vagrant up`
- `vagrant ssh-config` failing

Not to mention we have to manually build libssh as well as krb5.

Let's not mess with all that.

With this PR, we also don't /need/ to run centos8-vagrant.yml.  It can be used to RE-install vagrant but the separate playbook is no longer required.

Signed-off-by: David Galloway <dgallowa@redhat.com>