ceph-pull-requests-arm64: reenable test on arm64
since we now perform the "make check" test only on ubuntu focal, on
which all the dependencies are available on both aarch64 and amd64,
and the "make check" passes on this architecture also. let's test the
PRs on this architecture.

this change reverts 1e0b902baa33b3b972891779d2950e580bb18e63

Signed-off-by: Kefu Chai <kchai@redhat.com>