ceph-dashboard-cephadm-e2e-job: update workspace path
In the nightly runs we saw that the ceph_cluster.yml file which we used
to deploy kcli plans is not up-to-date with the master one. It cause
problems in the nightlies. We've narrowed down the problem is with the basedir
and and also with the path difference in the dependency script and the location of
ceph. So changing the workspace path to workspace/ceph.

Signed-off-by: Nizamudeen A <nia@redhat.com>