ceph-{,dev,dev-new} build: fix conditional for cephadm package test
https://github.com/ceph/ceph-build/pull/1919 added a conditional
test to check for cephadm package existence (so the code to extract
cephadm could work on older builds that did not have a cephadm package).
However, it used bash's [[ operator, and that doesn't expand file globs
in its arguments, so it effectively stopped creating the cephadm
binary altogether.

Switch to [, which doesn't have that problem.

Signed-off-by: Dan Mick <dmick@redhat.com>