ceph-dev-container-only: Create job
This job will allow users to build a container for https://quay.ceph.io/repository/ceph-ci/ceph?tab=tags given a branch name.  Should work for ceph.git and ceph-ci.git.

This is mostly useful for when packages and a repo successfully get built for a ceph-dev*build job but the container build fails.

Signed-off-by: David Galloway <dgallowa@redhat.com>