[ceph-windows-installer-build] Remove `FORCE` parameter
Remove the `FORCE` parameter from the `ceph-windows-installer-build`
Jenkins job, and use it by default.

This job builds the installer from source code, and it bundles
the Ceph binaries and the WNBD driver.

Therefore, the logic to skip the build if the latest installer commit
didn't change is not useful, because the installer might contain
different Ceph binaries and WNBD driver.