cephadm-dashboard-e2e: archiving takes too long to fail
If the provided pattern for eg: log/*.log is not found, archiving takes
way too longer to fail and that increased the job time to 1 hour on our
cephadm dashboard e2e jobs. Instead I added a pattern like log/** and
this way it immediately failed when there is no logs. (In cephadm
dashboard e2e, there
won't be any logs if the job is succesfull and the logs are saved only
when the job is failed)

Signed-off-by: Nizamudeen A <nia@redhat.com>