[wnbd-build] Remove `FORCE` parameter
Remove the `FORCE` parameter from the `wnbd-build` Jenkins job, and
use it by default.

The `wnbd` project is not updated that frequently, so the daily
build trigger mostly skips the Chacra upload. This is a problem
because when old Chacra artifacts are deleted, the `wnbd` build will
not be available in Chacra.

We need `wnbd` to always be available in Chacra, since the
Windows Teuthology tests will try to download it from there.