dashboard-deps: remove cortx from the apt-sources
dashboard e2e is constantly failing with this error for many days. So I
went for the simple solution here which is to remove the cortx from the
apt-sources. What I am not sure is the impact of this change. Will it
affect any other tests?

I also did some cleanups around cephadm e2e. Ideally this should be done
in the on_cleanup() function in our start-cluster.sh script in the ceph
repo (which I'll be doing), but for extra safety I am putting it here
too.

Signed-off-by: Nizamudeen A <nia@redhat.com>