ceph-windows: update test timeouts
The Python rbd-wnbd tests time out on stable branches. The reason
is that if the test script is missing, we're fetching it from
the main branch and then run the tests individually.

The issue is that while on the main branch the entire suite uses
a 30m timeout, on stable branches we're excuting each individual
test with a 5m timeout, which isn't enough for some of the tests
(e.g. the FIO one).

For this reason, we're going to increase the timeout.

While at it, we're pinning the commit id when fetching the test
script on older branches. That will allow us to move it or
refactor it.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>